third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
CONSOLE ERROR: Error with Suborigin header: Invalid character '@' in suborigin.
CONSOLE ERROR: Error with Suborigin header: Invalid character 'b' in suborigin policy. Suborigin policy options must start and end with a single quote.
CONSOLE ERROR: Error with Suborigin header: Invalid character 'F' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character 'F' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character '4' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character '-' in suborigin.
CONSOLE ERROR: Error with Suborigin header: Invalid character '-' in suborigin. First character must be a lower case alphabetic character.
CONSOLE ERROR: Error with Suborigin header: Invalid character '-' in suborigin.
This is a testharness.js-based test.
PASS Invalid suborigin names
Harness: the test ran to completion.
